Get Semáforo Epidemiológico Data from Mexico COVID-19 Response

Get Semáforo Epidemiológico Data from Mexico COVID-19 Response

semaforos_get(format = c("long", "wide"))

Arguments

format

Either "long" (the default), or "wide" to specify the format of the output data.frame

Value

if format = "long", a data.frame with 4 columns:

estado which contains the name of the state to which the semaforo in the row applies

year the four-digit year (%Y in strptime())

week the two-digit week, as specified in the ISO8601 (%W in strptime()

color the color of the semaforo, indicating the alert level

if format = "wide", a data.frame with estado as the first column and then one column for each week, with the column names indicating the week in the format %Y-W%W (e.g., "2021-W02")

Examples

sw <- semaforos_get(format = "wide")
head(sw)
#>                estado 2020-W22 2020-W23 2020-W24 2020-W25 2020-W26 2020-W27
#> 1      Aguascalientes     rojo     rojo  naranja  naranja  naranja  naranja
#> 2     Baja California     rojo     rojo     rojo     rojo     rojo     rojo
#> 3 Baja California Sur     rojo     rojo  naranja  naranja  naranja  naranja
#> 4            Campeche     rojo     rojo  naranja  naranja  naranja  naranja
#> 5             Chiapas     rojo     rojo     rojo     rojo  naranja     rojo
#> 6           Chihuahua     rojo     rojo  naranja  naranja  naranja  naranja
#>   2020-W28 2020-W29 2020-W30 2020-W31 2020-W32 2020-W33 2020-W34 2020-W35
#> 1  naranja  naranja  naranja  naranja  naranja     rojo     rojo  naranja
#> 2     rojo  naranja  naranja  naranja  naranja  naranja  naranja  naranja
#> 3  naranja     rojo     rojo     rojo     rojo     rojo     rojo  naranja
#> 4  naranja  naranja  naranja  naranja  naranja amarillo amarillo amarillo
#> 5     rojo  naranja  naranja  naranja  naranja  naranja  naranja amarillo
#> 6  naranja  naranja  naranja  naranja  naranja  naranja  naranja amarillo
#>   2020-W36 2020-W37 2020-W38 2020-W39 2020-W40 2020-W41 2020-W42 2020-W43
#> 1  naranja  naranja  naranja amarillo amarillo  naranja  naranja  naranja
#> 2  naranja  naranja  naranja amarillo amarillo amarillo amarillo  naranja
#> 3  naranja  naranja  naranja  naranja  naranja amarillo amarillo  naranja
#> 4 amarillo amarillo amarillo    verde    verde    verde    verde    verde
#> 5 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#> 6 amarillo amarillo amarillo amarillo amarillo  naranja  naranja     rojo
#>   2020-W44 2020-W45 2020-W46 2020-W47 2020-W48 2020-W49 2020-W50 2020-W51
#> 1  naranja  naranja  naranja  naranja  naranja  naranja  naranja  naranja
#> 2  naranja  naranja  naranja  naranja  naranja     rojo     rojo     rojo
#> 3  naranja  naranja  naranja amarillo amarillo  naranja  naranja  naranja
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5 amarillo amarillo amarillo    verde    verde    verde    verde    verde
#> 6     rojo     rojo     rojo     rojo     rojo  naranja  naranja  naranja
#>   2020-W52 2021-W01 2021-W02 2021-W03 2021-W04 2021-W05 2021-W06 2021-W07
#> 1  naranja amarillo amarillo  naranja  naranja  naranja  naranja  naranja
#> 2     rojo     rojo     rojo  naranja  naranja  naranja  naranja amarillo
#> 3  naranja  naranja  naranja  naranja  naranja  naranja  naranja  naranja
#> 4    verde    verde    verde    verde    verde amarillo amarillo amarillo
#> 5    verde    verde    verde amarillo amarillo amarillo amarillo    verde
#> 6  naranja  naranja  naranja amarillo amarillo  naranja  naranja amarillo
#>   2021-W08 2021-W09 2021-W10 2021-W11 2021-W12 2021-W13 2021-W14 2021-W15
#> 1  naranja  naranja amarillo amarillo amarillo amarillo amarillo amarillo
#> 2 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#> 3  naranja  naranja amarillo amarillo amarillo amarillo amarillo  naranja
#> 4 amarillo amarillo    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6 amarillo amarillo amarillo amarillo amarillo  naranja  naranja  naranja
#>   2021-W16 2021-W17 2021-W18 2021-W19 2021-W20 2021-W21 2021-W22 2021-W23
#> 1 amarillo amarillo amarillo amarillo amarillo    verde    verde    verde
#> 2 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#> 3  naranja  naranja  naranja  naranja amarillo amarillo amarillo  naranja
#> 4    verde    verde    verde    verde    verde amarillo amarillo amarillo
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6  naranja  naranja  naranja  naranja  naranja amarillo amarillo amarillo
#>   2021-W24 2021-W25 2021-W26 2021-W27 2021-W28 2021-W29 2021-W30 2021-W31
#> 1    verde    verde    verde    verde    verde    verde    verde    verde
#> 2 amarillo    verde    verde    verde    verde    verde amarillo amarillo
#> 3  naranja  naranja  naranja  naranja  naranja  naranja  naranja  naranja
#> 4 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#>   2021-W32 2021-W33 2021-W34 2021-W35 2021-W36 2021-W37 2021-W38 2021-W39
#> 1 amarillo amarillo  naranja  naranja  naranja  naranja amarillo amarillo
#> 2 amarillo amarillo amarillo amarillo amarillo amarillo amarillo amarillo
#> 3  naranja  naranja amarillo amarillo amarillo amarillo    verde    verde
#> 4 amarillo amarillo  naranja  naranja  naranja  naranja amarillo amarillo
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6 amarillo amarillo amarillo amarillo    verde    verde    verde    verde
#>   2021-W40 2021-W41 2021-W42 2021-W43 2021-W44 2021-W45 2021-W46 2021-W47
#> 1 amarillo amarillo amarillo amarillo amarillo amarillo    verde    verde
#> 2  naranja  naranja  naranja  naranja  naranja  naranja  naranja  naranja
#> 3    verde    verde    verde    verde    verde    verde    verde    verde
#> 4 amarillo amarillo amarillo amarillo    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6    verde    verde amarillo amarillo    verde    verde    verde    verde
#>   2021-W48 2021-W49 2021-W50 2021-W51 2021-W52 2022-W02 2022-W03 2022-W04
#> 1    verde    verde amarillo amarillo amarillo amarillo amarillo     rojo
#> 2 amarillo amarillo amarillo amarillo amarillo amarillo amarillo  naranja
#> 3 amarillo amarillo    verde    verde    verde  naranja  naranja  naranja
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6 amarillo amarillo amarillo amarillo amarillo  naranja  naranja  naranja
#>   2022-W05 2022-W06 2022-W07 2022-W08 2022-W09 2022-W10 2022-W11 2022-W12
#> 1     rojo  naranja  naranja amarillo amarillo    verde    verde    verde
#> 2  naranja  naranja  naranja    verde    verde    verde    verde    verde
#> 3  naranja amarillo amarillo    verde    verde    verde    verde    verde
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6  naranja  naranja  naranja amarillo amarillo    verde    verde    verde
#>   2022-W13 2022-W14 2022-W15 2022-W16 2022-W17 2022-W18 2022-W19 2022-W20
#> 1    verde    verde    verde    verde    verde    verde    verde    verde
#> 2    verde    verde    verde    verde    verde    verde    verde    verde
#> 3    verde    verde    verde    verde    verde    verde    verde    verde
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6    verde    verde    verde    verde    verde    verde    verde    verde
#>   2022-W21 2022-W22 2022-W23 2022-W24 2022-W25 2022-W26 2022-W27 2022-W28
#> 1    verde    verde    verde    verde    verde    verde    verde    verde
#> 2    verde    verde    verde    verde    verde    verde    verde    verde
#> 3    verde    verde    verde    verde    verde    verde    verde    verde
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6    verde    verde    verde    verde    verde    verde    verde    verde
#>   2022-W29 2022-W30 2022-W31 2022-W32 2022-W33 2022-W34 2022-W35 2022-W36
#> 1    verde    verde    verde    verde    verde    verde    verde    verde
#> 2    verde    verde    verde    verde    verde    verde    verde    verde
#> 3    verde    verde    verde    verde    verde    verde    verde    verde
#> 4    verde    verde    verde    verde    verde    verde    verde    verde
#> 5    verde    verde    verde    verde    verde    verde    verde    verde
#> 6    verde    verde    verde    verde    verde    verde    verde    verde
#>   2022-W37
#> 1    verde
#> 2    verde
#> 3    verde
#> 4    verde
#> 5    verde
#> 6    verde

sl <- semaforos_get(format = "long")
head(sl)
#>                estado year week color
#> 1      Aguascalientes 2020   22  rojo
#> 2     Baja California 2020   22  rojo
#> 3 Baja California Sur 2020   22  rojo
#> 4            Campeche 2020   22  rojo
#> 5             Chiapas 2020   22  rojo
#> 6           Chihuahua 2020   22  rojo